fix: Cliente Mapper and id_cliente column

+++ src/main/java/com/example/BancoMalvaderWeb/DTO/ContaDTO.java
@@ -7,7 +7,7 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ContaDTO {
-    private Double idConta;
+    private Long idConta;
     private String numeroConta;
     private String agencia;
     private double saldo;
